requirements.txt
filetype
Pillow
tweepy
requests[socks]
